corpcraft/aicrm_system/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Web Scraping
scrapy==2.11.0
selenium==4.15.2
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Anti-Scraping
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
selenium-stealth==1.0.6
requests-html==0.10.0

# Proxy & Rate Limiting
httpx[socks]==0.25.2
aiohttp-socks==0.8.4
ratelimit==2.2.1
slowapi==0.1.9

# CAPTCHA Solving
ddddocr==1.5.5
pytesseract==0.3.10
Pillow==10.1.0
2captcha-python==1.0.2

# OCR & Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
pyyaml==6.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
pymongo==4.6.0

# AI & ML
openai==1.3.7
anthropic==0.7.8
langchain==0.0.335
langchain-openai==0.0.2
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Task Queue
celery==5.3.4
kombu==5.3.4

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
